requirements.txt
asgiref==3.7.2
attrs==23.2.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0.1
django-background-tasks==1.2.8
django-easy-select2==1.5.8
django-filter==23.5
django-jalali==6.0.1
django-jalali-date==1.1.0
django-memcached==0.1.2
django-resized==1.0.2
djangoql==0.18.1
djangorestframework==3.14.0
drf_api_logger==1.1.15
et-xmlfile==1.1.0
idna==3.6
isodate==0.6.1
jalali==1.2
jdate==0.0.3
jdatetime==4.1.1
Khayyam==3.0.17
lxml==5.1.0
mysqlclient==2.2.1
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
pycparser==2.21
pymemcache==4.0.0
pyodbc==5.0.1
python-dateutil==2.8.2
python-memcached==1.62
pytz==2023.3.post1
pywin32==306
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
WMI==1.5.1
xlwt==1.3.0
zeep==4.2.1
